clean up mock dependencies
