+++ WireAuthentication/Sources/WireAuthenticationUI/Mocks/MockDependencies+FetchBackendConfigUseCaseFactory.swift
@@ -1,0 +1,50 @@
+//
+// Wire
+// Copyright (C) 2025 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+import WireAuthenticationAPI
+
+extension MockDependencies: FetchBackendConfigUseCaseFactory {
+
+    nonisolated
+    func fetchBackendConfigUseCase() -> any FetchBackendConfigUseCaseProtocol {
+        MockFetchBackendConfigUseCase()
+    }
+
+}
+
+struct MockFetchBackendConfigUseCase: FetchBackendConfigUseCaseProtocol {
+
+    func invoke(at configURL: URL) async throws -> BackendConfig {
+        BackendConfig(
+            title: "backend name",
+            endpoints: Endpoints(
+                backendURL: URL(string: "example")!,
+                backendWSURL: URL(string: "example")!,
+                blackListURL: URL(string: "example")!,
+                teamsURL: URL(string: "example")!,
+                accountsURL: URL(string: "example")!,
+                websiteURL: URL(string: "example")!,
+                countlyURL: URL(string: "example")!
+            ),
+            proxySettings: nil,
+            pinnedKeys: nil
+        )
+    }
+
+}
